# [api] Improve query performance