With the introduction of the virtual attributes, querying with
expanded resources and specific attributes (physical or virtual) has
been crawling terribly, i.e. querying ~300 vms with details slowed
down from ~1sec to ~22sec.

The issue was that we kept querying object reflections and virtual columns
with each attribute check !!. This fix now caches this information per class
and our queries are back to normal, down from ~22sec to ~1sec !!