@Fryguy 2015-02-25
I found an issue in the VirtualColumns stuff where we are not caching when I thought we used to.  This introduces an extra Hash merge into the mix everytime the method is called.

before caching:

```ruby
[1] pry(main)> Benchmark.ips { |x| x.report("ref") { VmVmware.reflections_with_virtual }; x.report("col") { VmVmware.virtual_columns_hash } }
Calculating -------------------------------------
                 ref     1.056k i/100ms
                 col   267.000  i/100ms
-------------------------------------------------
                 ref     16.561k (±31.4%) i/s -     68.640k
                 col      3.270k (±31.5%) i/s -     13.884k
```

after heavy caching (not sure if the right thing to do yet, but just trying to see numbers):

```ruby
[1] pry(main)> Benchmark.ips { |x| x.report("ref") { VmVmware.reflections_with_virtual }; x.report("col") { VmVmware.virtual_columns_hash } }
Calculating -------------------------------------
                 ref    36.152k i/100ms
                 col    33.598k i/100ms
-------------------------------------------------
                 ref      5.912M (± 8.3%) i/s -     29.283M
                 col      1.560M (± 8.6%) i/s -      7.761M
```

cc @tenderlove I'm going to make some changes to the virtual columns stuff, so wanted to give you a heads up.