@jrafanie 2015-02-25
Awesome find @abellotti... even if we close this and fix elsewhere, we'll fix the slow down :heart: