@abellotti 2015-02-25
I still need to do some caching on the rest api side in this PR, but for something else (from 1.0) which resulting in calling attr_virtual? for each attribute of a resource instead of each attribute of a requested attribute (i.e. 70+ calls per vm resource down to number of requested attributes).  With this fix including Jason's magic with reflections_with_virtual, we should have *much* better results. Thanks.