@tenderlove 2015-02-26
We should probably figure out where we're calling `reflections_with_virtual` and stop doing that.  Most of the ones I've seen are just checking to see if a reflection exists, and we can do that with (possibly) two hash lookups.  I think it's pretty rare that someone *actually* needs access to this hash.