@abellotti 2015-02-26
From what I can tell, both api and reporting/miq_expression use it, traversing reflections that can be queried, actual or virtual.