@tenderlove 2015-02-26
Well, it looks like we are only traversing this hash in two places: [here](https://github.com/ManageIQ/manageiq/blob/e02fb5258a5b6ea9ce5fead814e1544189777ae2/vmdb/app/models/mixins/ar_count_mixin.rb#L5) and [here](https://github.com/ManageIQ/manageiq/blob/e02fb5258a5b6ea9ce5fead814e1544189777ae2/vmdb/app/models/miq_expression.rb#L1491).  One of these calls is at boot time, so it shouldn't be a runtime bottleneck.  I'm not sure about the other one.

All other places that call `reflections_with_virtual` are just accessing the hash and could be replaced with `reflection_with_virtual`.