@tenderlove 2015-02-26
I sent #1905 to replace all the calls to `reflections_with_virtual` with `reflection_with_virtual` except the ones in this PR.