@abellotti 2015-02-26
Ok, additional enhancements this morning. Apparently, the query on attribute type was for each resource attribute even when asking for a subset, what did this mean:

With yesterday's fix, we got down querying 360 vms and the name attribute, down from 22 seconds back to 1.5 or so.  However, today I tried with 1000 VMs as a basis.  Querying 1000 vms, with the name attribute, would timeout the request, i.e. NetTimeout > 2minutes. Not a good thing.

WIth the fixes today, GET /api/vms?expand=resources&attributes=name&limit=1000 returns in 2.4 seconds!!

```
  Type of query                               Before                           Now
  GET 1000 w/ name attribute        Timeout (>2minutes)     2.4 seconds
  GET 500 w/ name attribute          32 seconds                    1.4 seconds
```

With Jason's virtual column caching fix, we'd be even better off.