@abellotti 2015-02-26
Just tested the same query above GET 1000 vms with name attribute with Jason's caching logic added to the virtual_* and *_with_virtual methods, the query now only take 1.2 seconds, down from 2.4 !!!!!  Thanks Jason.